+++ js/admin.js
@@ -18,6 +18,9 @@ app.initializers.add('server-side-highlight', function(app) {
       type: 'select',
       options: themes,
     })
+    .registerSetting(function () {
+      return preview.bind(this)('light');
+    })
     .registerSetting({
       setting: 'club-1-server-side-highlight.dark_theme_bg_color',
       label: app.translator.trans('club-1-server-side-highlight.admin.dark_theme_bg_color'),
@@ -33,7 +36,22 @@ app.initializers.add('server-side-highlight', function(app) {
       label: app.translator.trans('club-1-server-side-highlight.admin.dark_theme_highlight_theme'),
       type: 'select',
       options: themes,
+    })
+    .registerSetting(function () {
+      return preview.bind(this)('dark');
     });
 });
+
+function preview(mode) {
+  let bg = encodeURIComponent(this.setting(`club-1-server-side-highlight.${mode}_theme_bg_color`)());
+  let text = encodeURIComponent(this.setting(`club-1-server-side-highlight.${mode}_theme_text_color`)());
+  let theme = encodeURIComponent(this.setting(`club-1-server-side-highlight.${mode}_theme_highlight_theme`)());
+  return m('div.Form-group', m('iframe', {
+    src: app.route('dashboard') + `admin/highlight-preview?bg=${bg}&text=${text}&theme=${theme}`,
+    width: 400,
+    height: 220,
+    style: 'border: none; border-radius: var(--border-radius)',
+  }));
+}
 
  module.exports = {}